Revert "Merge branch 'feature/mongodb-implementation' into develop"
This reverts commit d0b9bb92f1ded8a94954150b508cbfa0ab605886, reversing
changes made to 150fc9f5917618f2600dc58735ca767c55eb6d37.

+++ TiendaLaSebastiana/src/BusinessLogic/Reporte.java
@@ -61,12 +61,12 @@ public class Reporte {
         return ventasFiltradas;
     }
     
-        public Map<Long, Integer> filtrarStack() {
-        Map<Long, Integer> cantidadesFiltradas = new HashMap<>();
+        public Map<Long, Double> filtrarStack() {
+        Map<Long, Double> cantidadesFiltradas = new HashMap<>();
 
-        for (Map.Entry<Long, Integer> acceso : caja.getInventario().getCantidades().entrySet()) {
+        for (Map.Entry<Long, Double> acceso : caja.getInventario().getCantidades().entrySet()) {
             Long productoId = acceso.getKey();
-            Integer cantidad = acceso.getValue();
+            Double cantidad = acceso.getValue();
             if (cantidad < 5 || cantidad > 20) {
                 cantidadesFiltradas.put(productoId, cantidad);
             }
@@ -75,7 +75,7 @@ public class Reporte {
         return cantidadesFiltradas;
     }
     
-    public ArrayList<Producto> filtrarProductos(Map<Long, Integer> cantidadesFiltradas) {
+    public ArrayList<Producto> filtrarProductos(Map<Long, Double> cantidadesFiltradas) {
         ArrayList<Producto> productosFiltrados = new ArrayList<>();
 
         for (Producto producto : caja.getInventario().getProductos()) {
